requirements.txt
aiohttp==3.8.4
flake8==6.0.0
marshmallow==3.19.0
marshmallow-dataclass==8.5.11
PyYAML==6.0
aio-pika==9.0.2
SQLAlchemy==1.4.40
alembic==1.8.1
asyncpg==0.27.0
pytest==7.2.1
pytest-asyncio==0.20.3
pytest-mock==3.10.0
pytest-aiohttp==1.0.4

anyio
pytest-tornasync
pytest-trio
